Refactor `printPropertyKey`
