+++ src/language-js/print/property.js
@@ -1,16 +1,68 @@
+import isEs5IdentifierName from "@prettier/is-es5-identifier-name";
+
 import { printComments } from "../../main/comments/print.js";
 import printNumber from "../../utils/print-number.js";
 import printString from "../../utils/print-string.js";
-import {
-  isNumericLiteral,
-  isSimpleNumber,
-  isStringLiteral,
-  isStringPropSafeToUnquote,
-  rawText,
-} from "../utils/index.js";
+import { isNumericLiteral, isStringLiteral, rawText } from "../utils/index.js";
 import { printAssignment } from "./assignment.js";
 
 const needsQuoteProps = new WeakMap();
+
+// Matches “simple” numbers like `123` and `2.5` but not `1_000`, `1e+100` or `0b10`.
+function isSimpleNumber(numberString) {
+  return /^(?:\d+|\d+\.\d+)$/.test(numberString);
+}
+
+// Note: Quoting/unquoting numbers in TypeScript is not safe.
+//
+// let a = { 1: 1, 2: 2 }
+// let b = { '1': 1, '2': 2 }
+//
+// declare let aa: keyof typeof a;
+// declare let bb: keyof typeof b;
+//
+// aa = bb;
+// ^^
+// Type '"1" | "2"' is not assignable to type '1 | 2'.
+//   Type '"1"' is not assignable to type '1 | 2'.(2322)
+//
+// And in Flow, you get:
+//
+// const x = {
+//   0: 1
+//   ^ Non-string literal property keys not supported. [unsupported-syntax]
+// }
+//
+// Angular does not support unquoted numbers in expressions.
+//
+// So we play it safe and only unquote numbers for the JavaScript parsers.
+// (Vue supports unquoted numbers in expressions, but let’s keep it simple.)
+//
+// Identifiers can be unquoted in more circumstances, though.
+function isStringPropSafeToUnquote(node, options) {
+  return (
+    options.parser !== "json" &&
+    options.parser !== "jsonc" &&
+    isStringLiteral(node.key) &&
+    rawText(node.key).slice(1, -1) === node.key.value &&
+    ((isEs5IdentifierName(node.key.value) &&
+      // With `--strictPropertyInitialization`, TS treats properties with quoted names differently than unquoted ones.
+      // See https://github.com/microsoft/TypeScript/pull/20075
+      !(
+        (options.parser === "babel-ts" && node.type === "ClassProperty") ||
+        (options.parser === "typescript" && node.type === "PropertyDefinition")
+      )) ||
+      (isSimpleNumber(node.key.value) &&
+        String(Number(node.key.value)) === node.key.value &&
+        // TODO[@fisker]: `ImportAttribute` should not care about parser
+        node.type !== "ImportAttribute" &&
+        (options.parser === "babel" ||
+          options.parser === "acorn" ||
+          options.parser === "espree" ||
+          options.parser === "meriyah" ||
+          options.parser === "__babel_estree")))
+  );
+}
 
 function printPropertyKey(path, options, print) {
   const { node } = path;
